requirements.txt
amqp==2.6.1
argon2-cffi==20.1.0
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
coverage==5.4
Django==3.1.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-uuslug==1.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
flower==0.9.7
humanize==3.2.0
idna==2.10
importlib-metadata==3.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==4.6.11
MarkupSafe==1.1.1
packaging==20.9
Pillow==8.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.9
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.0
pyparsing==2.4.7
python-slugify==4.0.1
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
tornado==6.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
zipp==3.4.0
